keep plugin sources with tag back to github.com/HugoBlox/
Update go.mod
* remove pseudo-version
invalid pseudo-version: preceding tag (v1.1.1) not found
This error means Go is trying to resolve a pseudo-version like v1.1.2-0.20231209203044-d31adfedd40b, but it can't find the preceding tag v1.1.1 in the Git history of the module blox-plugin-netlify. Pseudo-versions rely on semantic version tags to determine their place in the version timeline, and if the tag is missing, Go can't validate the version.

Update go.mod